+++ fixpoint_cpp_simulator/fixpoint_cpp_simulator/utils.h
@@ -17,10 +17,10 @@ struct param
 param calc_param(int m, int n);
 
 template <class T>
-void pickarray(T const din[], T dout[], size_t start, size_t step, size_t length)
+void pickarray(T const din[], T dout[], int start, int step, int length)
 {
-	size_t j = 0;
-	for (size_t i = start; i < length; i += step , j++)
+	int j = 0;
+	for (int i = start; i < length; i += step , j++)
 	{
 		dout[j] = din[i];
 	}
